requirements.txt
requests==2.33.1pandas==3.0.2python-dotenv==1.2.2matplotlib==3.10.9seaborn==0.13.2numpy==2.4.4